+++ src/doitoml/errors.py
@@ -31,6 +31,11 @@ class UnsafePathError(ConfigError):
     """An error related to unsafe paths."""
 
 
+class SkipError(ConfigError):
+
+    """An error related to an ambiguous skip."""
+
+
 class NoConfigError(ConfigError):
 
     """An error when no configuration at all is found."""