+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - adds `doitoml[test]`
 - adds `safe_paths` to the Python API
   - arbitrary paths in _shell_ actions are uncheckable, however
+- expands `skip` to allow custom skippers like `any`, `all`, `not`, `exists`, `platform`
 
 ## 0.1.1